langgraph_agent/requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
langgraph
langgraph-checkpoint-sqlite

absl-py
